hw10: apk добавлен в гит, исходники распределены по пакетам
