--- app/src/main/java/com/example/notes/NotesAdapter.java
@@ -1,43 +1,0 @@
-package com.example.notes;
-
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class NotesAdapter extends RecyclerView.Adapter<NoteViewHolder> {
-    private List<NoteEntity> list = new ArrayList<>();
-    private OnItemClickListener onItemClickListener;
-
-    public void setList(List<NoteEntity> list) {
-        this.list = list;
-        notifyDataSetChanged();
-    }
-
-    interface OnItemClickListener {
-        void onItemClick(NoteEntity note);
-    }
-
-    public void setOnItemClickListener(OnItemClickListener onItemClickListener) {
-        this.onItemClickListener = onItemClickListener;
-    }
-
-    @NonNull
-    @Override
-    public NoteViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        return new NoteViewHolder(parent, onItemClickListener);
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull NoteViewHolder holder, int position) {
-        holder.bind(list.get(position));
-    }
-
-    @Override
-    public int getItemCount() {
-        return list.size();
-    }
-}
