--- app/src/main/java/com/example/notes/NotesRepo.java
@@ -1,16 +1,0 @@
-package com.example.notes;
-
-import java.util.ArrayList;
-
-public interface NotesRepo {
-    void addNote(NoteEntity note);
-    boolean updateNote(NoteEntity newNote);
-    void deleteNote(NoteEntity note);
-    ArrayList<NoteEntity> getNotes();
-
-    void setListener(Notifier notifier);
-
-    public interface Notifier {
-        void updateRepo();
-    }
-}
